made the back button independent of the tabs
